feat(starlight-karma-bot): import legacy database automatically on startup
